DetailsController API

+++ Vehicles.API/Models/DetailViewModel.cs
@@ -13,12 +13,12 @@ namespace Vehicles.API.Models
         [Display(Name = "Precio Mano de Obra")]
         [DisplayFormat(DataFormatString = "{0:C2}")]
         [Required(ErrorMessage = "El campo {0} es obligatorio.")]
-        public decimal LaborPrice { get; set; }
+        public int LaborPrice { get; set; }
 
         [Display(Name = "Precio Repuestos")]
         [DisplayFormat(DataFormatString = "{0:C2}")]
         [Required(ErrorMessage = "El campo {0} es obligatorio.")]
-        public decimal SparePartsPrice { get; set; }
+        public int SparePartsPrice { get; set; }
 
         [Display(Name = "Observación")]
         [DataType(DataType.MultilineText)]
